Added mechanisms to the app (new mechanism, show mechanism in the tree, rename/copy mechanism, etc).
Set the title of the workspace name modal bases on what is being done (new workspace, rename workspace, copy workspace).
Set the title of the module name modal bases on what is being done (new workspace, rename workspace, copy workspace).

Fixed a problem where the module type wasn't being passed to the python generator when it generates the python code for the UI.

When a new module is selected and we are updating the blocks editor area, close the toolbox flyout if it is open.

When saving a module to storage, include the module type.
Improved the code that writes and reads modules (Workspaces, Mechanisms, and OpModes) to/from storage.