In mrc_call_python_function.ts: Added INSTANCE_MECHANISM to FunctionKind enum to represent calling a mechanism instance method from the robot or from an OpMode. In renameMethodCaller, added code to update a block calling a component instance method if the component name is changed, and to update a block calling a mechanism instance method if the mechanism name or method name/signature is changed. Added functions addInstanceMechanismBlocks and createInstanceMechanismBlock.
In mrc_component.ts:
Added validator to the name field so we can call renameMethodCallers when the user changes the component name.
Removed hideParams and mrcHideParams. Instead use module type to decide whether to hide parameters.
Replaced getNewPort method with getArgName method, which uses both the component name and the arg name so the users knows what kind of ports (or other arguments) they are passing in.

In mrc_mechanism.ts:
Added validator to the name field so we can call renameMethodCallers when the user changes the mechanism name.

In editor.ts:
Added getMechanism method to get the Mechanism that correponds to the given MechanismInRobot.

In common_storage.ts:
Added some comments on className fields.

In hardware_category.ts:
Added code to getRobotMechanismsCategory to add a toolbox category with the methods for each mechanism that was added to the Robot.

In methods_category.ts:
Removed duplicate 'register_event' value.