Add support for mechanism event handlers (#199)
* Make contents optional in Category class because custom categories don't have contents.

* Convert getCategory (in event_category.ts and methods_category.ts) from arrow functions to function declarations.
See https://google.github.io/styleguide/tsguide.html#function-declarations

In mrc_call_python_function.ts:
Removed unnecessary semicolon.
Changed "...method on a mechanism..." to "...method in a mechanism...".

* Renamed getRobotEventsCategory to getRobotEventHandlersCategory.
Renamed RobotEventsCategory to RobotEventHandlersCategory.

In editor.ts, updated code that parses the module content text for the robot and mechanisms to use ? : instead of if/else.

* Updated mrc_event_handler.ts to support mechanism events.

* Moved CUSTOM_CATEGORY_EVENT_HANDLERS_ROBOT, getRobotEventHandlersCategory, and RobotEventHandlersCategory from hardware_category.ts to event_handlers_category.ts.

* Modified event_handlers_category.ts to work for robot events and mechanism events.

Use the block id of the mrc_event blocks to identify which event handlers are already on the workspace.

* Update event handler blocks if the user renames a mechanism.

* Update python generator for registerering event handlers.

* Create functions for registering custom categories.

Changed custom values to begin with MRC_.