Pr private components in mechanism (#221)
* Add dist to .gitignore

* Add private components to mechanism

* Preface components inside mechanism

* Add tooltips

* Fix false warning for calling methods on components in mechanisms

* removed unused import

* Updated mechanism and robot module start json.

* Added comment specifying that it needs to call getAllComponentsFromMechanism so it can create the port parameters.

* Updated mrc_mechanism_component_holder.ts
Since the mrc_mechanism_component_holder block can't change once it has been
placed on the workspace, the code can be simplified.

* Updated mrc_call_python_function.ts to handle mechanism component methods.
The mechanismId (and mrcMechanismId) is set to the id of the mechanism so if the mechanism is renamed (in the robot), we update it.

Added mrcComponentNames and mrcMapComponentNameToId. These are filled in during mrcOnLoad and are used to populate the component name dropdown field.

* Set current version to 0.0.2.
Added privateComponents to module content.
We can use this to get the private components from a mechanism without having to
use a headless blockly workspace to process the blocks.
Added code to upgrade projects to version 0.0.2, which involves updating the
mrc_mechanism_component_holder block in the robot and adding privateComponents
in module content.

* Added code to editor to get the private components and pass them to
storageModuleContent.makeModuleContentText.

Made some methods private where possible.

Modified getAllComponentsFromMechanism so it gets the private components from a
mechanism's module content without having to use a headless blockly workspace to
process the mechanism's blocks.

* Added comment in getRobotMechanismsCategory specifying that it needs to call
getAllComponentsFromMechanism so it can create the port parameters.

Simplified getComponentsCategory to call getAllComponentsFromWorkspace for both robot and mechanism modules and added comments explaining why.

---------

Co-authored-by: Liz Looney <lizlooney@gmail.com>