Show button instead of event handler block in toolbox if the event handler has already been used (#222)
* Added class Button to toolbox/items.ts.
Added support for making a category initially expanded.
Cleaned up other classes in toolbox/items.ts.

Modified getHardwareCategory in toolbox/hardware_category.ts so the Hardware or Robot category are initially expanded.

Modified getToolboxJSON in toolbox/toolbox.ts so it prevent a problem with blockly choking on our Category class.

Modified code that puts event handler blocks in the toolbox so that if an event handler is already on the workspace, a button for that event is put in the toolbox. If the user hits the button, an info message tells the user that the event handler is already on the workspace.

* Use Blockly.Msg.When instead of 'when'.
If the blockly workspace is RTL, reverse the words in the button text.

* Generate the css classes for event handler buttons with the correct color for each theme.