In editor.ts: (#262)
Added code to onChangeAfterLoading to handle MOVE events. If the block being moved has a mrcOnMove method, call it.

In mrc_event.ts:
Added constants INPUT_TITLE, FIELD_PARAM_PREFIX, and WARNING_ID_NOT_IN_HOLDER.
Removed call to ChangeFramework.registerCallback.
Removed onBlockChanged.
Added mrcHasWarning to EventMixin.
Added mrcOnLoad so that the block is checked when the block is loaded.
Added mrcOnMove so that the block is checked when the block is moved.