Change placeholder python code (OpMode, etc) to match Peter's API (#353)
* Change placeholder python code (OpMode, etc) to match Peter's API

* Updated generate_json.py and regenerated json files.

* Updated json used when creating a new mechanism or opmode.

* Added upgrade code to rename 'loop' to 'Periodic' in opmode modules.
Added upgrade code to rename 'update' to 'opmode_periodic' in mechanism modules.

* Make mrc_component blocks have a tooltip.
Removed staticFunctionName and mrcStaticFunctionName.

* Added a comment explaining some code in checkOpMode.

* Always generate the define_hardware method for a robot.

* Updated constants for class and method names.

* Update extended_python_generator.ts:
Changed code that generates the type decorator to use the full decorator name (including the module blocks_base_class) so the Teleop decorator doesn't collide with the Teleop class.
Generate code that was in robot_base.py, but is not in wpilib.OpModeRobot.
Generate code that was in opmode.py, but is not in wpilib.PeriodicOpMode.