Added a few more components, including some that have non-port arguments (#368)
* Changes to python_tools/python_util.py:
Fixed isEnum.
Removed getComponentPortNameAndType.

Changes to python_tools/json_util.py:
Generate JSON componentArgs for component constructors, instead of componentPortName and componentPortType.
Added _createArgData.
Added _processComponent.

* Moved PortType and related functions from module_content.ts to python_json_types.ts

Added preupgrade_009_to_0010 to change Component.ports to Component.args.

* Added valueForComponentArgInput to value.ts. It creates blocks to plug into component input sockets.

* Changed blocks and the python generator where it used to assume that all components have a port argument and no other arguments.

Fixed the code that generates the robot when it creates a mechanism, passing one tuple of args for each component in the mechanism.

Fixed the code that generates a mechanism's __init__ and define_hardware methods. The __init__ and define_hardware methods take one parameter (which is is a tuple) for each component in the mechanism. The define_hardware method uses the * operator to unpack the elements of the tuple and pass each element as a separate positional argument to the component constructor.

* Increment the current version to 0.0.10.
Upgrade old projects so that Components have args instead of ports.